+++ weblate/machinery/openai.py
@@ -4,8 +4,9 @@
 
 from __future__ import annotations
 
+from collections import defaultdict
 from itertools import chain
-from typing import TYPE_CHECKING, Literal
+from typing import TYPE_CHECKING, Literal, overload
 
 from django.core.cache import cache
 
@@ -39,6 +40,11 @@ SEPARATOR = "\n==WEBLATE_PART==\n"
 SEPARATOR_PROMPT = f"""
 You receive an input as strings separated by {SEPARATOR} and
 your answer separates strings by {SEPARATOR}.
+"""
+REPHRASE_PROMPT = f"""
+You receive an input as the source and existing translation strings separated
+by {SEPARATOR} and you answer three rephrased translation strings separated by
+{SEPARATOR}.
 """
 GLOSSARY_PROMPT = """
 Use the following glossary during the translation:
@@ -98,19 +104,33 @@ class OpenAITranslation(BatchMachineTranslation):
             text = f"{text}."
         return text
 
-    def get_prompt(
-        self, source_language: str, target_language: str, texts: list[str], units: list
+    def _get_prompt(
+        self,
+        source_language: str,
+        target_language: str,
+        texts: list[str],
+        units: list[Unit | None],
+        *,
+        rephrase: bool = False,
     ) -> str:
         glossary = ""
         if any(units):
             glossary = render_glossary_units_tsv(
                 chain.from_iterable(
-                    get_glossary_terms(unit, include_variants=False) for unit in units
+                    get_glossary_terms(unit, include_variants=False)
+                    for unit in units
+                    if unit is not None
                 )
             )
             if glossary:
                 glossary = GLOSSARY_PROMPT.format(glossary)
-        separator = SEPARATOR_PROMPT if len(units) > 1 else ""
+
+        separator = ""
+        if rephrase:
+            separator = REPHRASE_PROMPT
+        elif len(units) > 1:
+            separator = SEPARATOR_PROMPT
+
         placeables = ""
         if any(self.replacement_start in text for text in texts):
             placeables = PLACEABLES_PROMPT.format(
@@ -128,6 +148,9 @@ class OpenAITranslation(BatchMachineTranslation):
             placeables=placeables,
         )
 
+    def _can_rephrase(self, unit: Unit | None) -> bool:
+        return unit is not None and unit.translated and not unit.readonly
+
     def download_multiple_translations(
         self,
         source,
@@ -136,17 +159,76 @@ class OpenAITranslation(BatchMachineTranslation):
         user=None,
         threshold: int = 75,
     ) -> DownloadMultipleTranslations:
+        rephrase: list[tuple[str, Unit]] = []
+        texts: list[str] = []
+        units: list[Unit | None] = []
+
+        # Separate rephrasing and new translations
+        for text, unit in sources:
+            if self._can_rephrase(unit):
+                rephrase.append((text, unit))
+            else:
+                texts.append(text)
+                units.append(unit)
+
+        # Collect results
+        result: DownloadMultipleTranslations = defaultdict(list)
+
+        # Fetch rephrasing each string separately
+        if rephrase:
+            for text, unit in rephrase:
+                self._download(result, source, language, [text], [unit], rephrase=True)
+
+        # Fetch translations in batch
+        if texts:
+            self._download(result, source, language, texts, units)
+
+        return result
+
+    @overload
+    def _download(
+        self,
+        result: DownloadMultipleTranslations,
+        source,
+        language,
+        texts: list[str],
+        units: list[Unit],
+        *,
+        rephrase: Literal[True],
+    ): ...
+    @overload
+    def _download(
+        self,
+        result: DownloadMultipleTranslations,
+        source,
+        language,
+        texts: list[str],
+        units: list[Unit | None],
+    ): ...
+    def _download(
+        self,
+        result: DownloadMultipleTranslations,
+        source,
+        language,
+        texts,
+        units,
+        *,
+        rephrase=False,
+    ):
         from openai.types.chat import (
             ChatCompletionSystemMessageParam,
             ChatCompletionUserMessageParam,
         )
 
-        texts = [text for text, _unit in sources]
-        units = [unit for _text, unit in sources]
-        prompt = self.get_prompt(source, language, texts, units)
+        prompt = self._get_prompt(source, language, texts, units, rephrase=rephrase)
         messages = [
             ChatCompletionSystemMessageParam(role="system", content=prompt),
-            ChatCompletionUserMessageParam(role="user", content=SEPARATOR.join(texts)),
+            ChatCompletionUserMessageParam(
+                role="user",
+                content=SEPARATOR.join(
+                    texts if not rephrase else [*texts, units[0].target]
+                ),
+            ),
         ]
 
         response = self.client.chat.completions.create(
@@ -156,8 +238,6 @@ class OpenAITranslation(BatchMachineTranslation):
             frequency_penalty=0,
             presence_penalty=0,
         )
-
-        result: DownloadMultipleTranslations = {}
 
         translations_string = response.choices[0].message.content
         if translations_string is None:
@@ -169,7 +249,7 @@ class OpenAITranslation(BatchMachineTranslation):
             raise MachineTranslationError("Blank assistant reply")
 
         translations = translations_string.split(SEPARATOR)
-        if len(translations) != len(texts):
+        if not rephrase and len(translations) != len(texts):
             self.report_error(
                 "Failed to parse assistant reply",
                 extra_log=translations_string,
@@ -179,20 +259,13 @@ class OpenAITranslation(BatchMachineTranslation):
                 f"Could not parse assistant reply, expected={len(texts)}, received={len(translations)}"
             )
 
-        for index, text in enumerate(texts):
-            # Extract the assistant's reply from the response
-            try:
-                translation = translations[index]
-            except IndexError:
-                self.report_error("Missing assistant reply")
-                continue
-
-            result[text] = [
+        for index, translation in enumerate(translations):
+            text = texts[index if not rephrase else 0]
+            result[text].append(
                 {
                     "text": translation,
                     "quality": self.max_score,
                     "service": self.name,
                     "source": text,
                 }
-            ]
-        return result
+            )